requirements.txt
openai>=1.16.0
requests
beautifulsoup4
markdownify
python-dotenv
